feat: converted to maven
