+++ Backendp3/src/main/java/Events/FullyBooked.java
@@ -1,0 +1,5 @@
+package Events;
+
+public class FullyBooked {
+
+}
