+++ Backendp3/src/main/java/Events/Published.java
@@ -1,0 +1,4 @@
+package Events;
+
+public class Published {
+}
