+++ Backendp3/src/main/java/Events/Removed.java
@@ -1,0 +1,7 @@
+package Events;
+//import com.fasterxml.jackson.core.type.TypeReference;
+//import com.fasterxml.jackson.databind.ObjectMapper;
+
+public class Removed {
+
+}
