--- Backendp3/src/main/java/Events/participantCancelled.java
@@ -1,4 +1,0 @@
-package Events;
-
-public class participantCancelled {
-}
