--- Backendp3/src/main/java/Events/published.java
@@ -1,4 +1,0 @@
-package Events;
-
-public class published {
-}
